// src/auth/permissions/permissions.enum.ts
export enum Permission {
  // Companies
  COMPANIES_READ = 'companies.read',
  COMPANIES_CREATE = 'companies.create',
  COMPANIES_UPDATE = 'companies.update',
  COMPANIES_DELETE = 'companies.delete',

  // Roles
  ROLES_READ = 'roles.read',
  ROLES_CREATE = 'roles.create',
  ROLES_UPDATE = 'roles.update',
  ROLES_DELETE = 'roles.delete',

  // Users
  USERS_READ = 'users.read',
  USERS_CREATE = 'users.create',
  USERS_UPDATE = 'users.update',
  USERS_DELETE = 'users.delete',

  // Customers
  CUSTOMERS_READ = 'customers.read',
  CUSTOMERS_CREATE = 'customers.create',
  CUSTOMERS_UPDATE = 'customers.update',
  CUSTOMERS_DELETE = 'customers.delete',

  // Books
  BOOKS_READ = 'books.read',
  BOOKS_CREATE = 'books.create',
  BOOKS_UPDATE = 'books.update',
  BOOKS_DELETE = 'books.delete',

  // Shipments
  SHIPMENTS_READ = 'shipments.read',
  SHIPMENTS_CREATE = 'shipments.create',
  SHIPMENTS_UPDATE = 'shipments.update',
  SHIPMENTS_DELETE = 'shipments.delete',

  // Payments
  PAYMENTS_READ = 'payments.read',
  PAYMENTS_CREATE = 'payments.create',
  PAYMENTS_UPDATE = 'payments.update',
  PAYMENTS_DELETE = 'payments.delete',
}